+++ src/Tablesample/Tablesample.php
@@ -19,7 +19,7 @@ final readonly class Tablesample
         private TablesampleMethod $tablesampleMethod,
         private float $percentage,
     ) {
-        if (0.00 > $percentage || 100.00 < $percentage) {
+        if (0.00 >= $percentage || 100.00 <= $percentage) {
             throw new TablesampleWalkerException('A percentage must be between 0 and 100');
         }
     }