+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 CHANGELOG
 =========
-## Unreleased
 2.4
----
+-----
+* Add Infection
+* Fix `TABLESAMPLE` walker: throw `TablesampleWalkerException` when hint is missing an entity name instead of silent crash
+* Ensure 0 and 100 are not allowed in `TABLESAMPLE`
+* Fix `TABLESAMPLE` tests
+* `WITH TIES` walker now uses own finalizer
 
 2.3
 ---